ETR-61 Feat: 로그아웃 컴포넌트 분리
