+++ src/pages/Mypage/Logout.tsx
@@ -1,0 +1,56 @@
+import { useState } from "react";
+import { useDispatch } from "react-redux";
+import { logout } from "../../features/auth/authSlice";
+import ConfirmModal from "../../components/Modal/ConfirmModal";
+
+
+function Logout({ onLogoutComplete }: { onLogoutComplete: () => void }) {
+  
+  const [showConfirmModal, setShowConfirmModal] = useState(false);
+  const dispatch = useDispatch();
+
+  const handleLogout = async () => {
+    const refreshToken = sessionStorage.getItem("refreshToken");
+
+    try {
+        // 백엔드에 로그아웃 요청 
+        const res = await fetch("http://localhost:8080/api/auth/logout", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        credentials: "include", // 쿠키 포함
+        body: JSON.stringify({ refreshToken }), // sessionStorage에서 꺼낸 값
+      });
+      const result = await res.json();
+      console.log(result.data)
+    } catch (err) {
+      console.error("서버 로그아웃 실패", err);
+      // 실패해도 클라이언트에서 로그아웃은 계속 진행
+    }
+
+    // 클라이언트 상태 정리
+    dispatch(logout());
+    sessionStorage.removeItem("accessToken");
+    sessionStorage.removeItem("refreshToken");
+    onLogoutComplete();
+  };
+
+
+  return (
+    <>
+      <button className="text-sm text-gray-400 mt-5" onClick={() => setShowConfirmModal(true)}>
+        로그아웃
+      </button>
+
+      {showConfirmModal && (
+        <ConfirmModal
+          message="로그아웃 하시겠습니까?"
+          buttonmessage="로그아웃"
+          onConfirm={handleLogout}
+          onCancel={() => setShowConfirmModal(false)}
+        />
+      )}
+    </>
+  );
+}
+
+export default Logout;
